Allow bookie garbage collection to be triggered manually from tests
The current gc tests rely on waiting on a timeout for gc to run. It's
never certain whether it has run or not or if it's still running. This
sucks. This patch allows tests to trigger a gc run and gives the client
a future to know when it has completed. I haven't changed the gc
algorithm, but now it runs in a scheduled executor rather than as a
Thread.

Change-Id: I1cfd701ba53cd73af55e76025ff467cb5dce34ac